Created a Blueberries Website using Breadcrumbs
FreshDeals: Blueberries
A site for buying groceries. 

Added breadcrumbs to this site, Here users will have a better feeling for what the site provides and the optional attributes of the product they are purchasing.

I focused on the layout of this page and implemented the breadcrumbs via HTML and CSS. The links to the breadcrumbs will not function.