Created Pinpoint E-Commerce
In this project, I have used my gained knowledge within the Color for UI lesson and applied it to an e-commerce checkout page. In the grayscale website, I applied colours to various elements and sections within the design to ensure users understand the error and active states and primary calls-to-action.