Created a Off-Platform Project: Tea Cozy
Project: Tea Cozy
In this project, I created a fictional tea shop website 

it Included design spec and image assets which helped me along the way. A design spec is an image of a web page outlined with its CSS properties and values. A designer usually creates these as a source of instructions for a web developer.